Initial checkin
SonarI2C is a library for Arduino which permits multiple cheap
ultrasonic sensors to be addressed over I2C. It relies on an external
pin expander and logic chip to address the sensors and multiplex the
response pulses to a hardware interrupt pin.

Initial checkin breadboard tested to 10 HC-SR04 units - not yet used in
anger.